+++ fiddles/node/fiddle-0031-SpawnProcess/app.js
@@ -1,0 +1,9 @@
+const spawn = require('child_process').spawn,
+  clear = spawn('clear'),
+  reflect = spawn('cat',['app.js']);
+
+clear.stdout.pipe(process.stdout);
+reflect.stdout.pipe(process.stdout);
+
+
+